refactor: 인터페이스 정리
