+++ src/api/til/type.ts
@@ -1,4 +1,4 @@
-import type { CommonResponse, Step, Comment, Til } from '@/api/type';
+import type { CommonResponse, Step, Comment, Til, MemberTil } from '@/api/type';
 
 // Til 요청
 
@@ -14,7 +14,6 @@ export interface PatchTilsRequest {
 
 export interface SubmitTilsRequest extends PatchTilsRequest {}
 
-// postComment
 export interface PostCommentsRequest {
   roadmapId: number;
   stepId: number;
@@ -22,28 +21,17 @@ export interface PostCommentsRequest {
   content: string;
 }
 
-// patchComment
 export interface PatchCommentsRequest {
   content: string;
 }
+
+// Til 응답
 
 export interface GetStepTilsResponse extends CommonResponse {
   result: {
     members: MemberTil[];
   };
 }
-
-export interface MemberTil {
-  tilId: number | null;
-  userId: number;
-  name: string;
-  image: string;
-  content: string | null;
-  submitDate: string | null;
-  commentNum: number | null;
-}
-
-// Til 응답
 
 export interface GetTilsResponse extends CommonResponse {
   result: {
